Move files from src/shared
Move files from `src/shared` to `core`
