+++ vimeo-uploader-lambda/Dockerfile
@@ -4,8 +4,7 @@ COPY requirements.txt  .
 RUN pip3 install -r requirements.txt --target "${LAMBDA_TASK_ROOT}"
 
 COPY app.py ${LAMBDA_TASK_ROOT}
-COPY src/shared ${LAMBDA_TASK_ROOT}/src/shared
-COPY src/generated ${LAMBDA_TASK_ROOT}/src/generated
+COPY core ${LAMBDA_TASK_ROOT}/core
 COPY ffmpeg-5.1.1-arm64-static /usr/local/bin/ffmpeg
 
 ENV PATH "${PATH}:/usr/local/bin/ffmpeg"
