+++ vimeo-uploader-lambda/core/exceptions.py
